refactor: Split app to agent and server
To simplify addition of new application.
