--- internal/app/docs.go
@@ -1,3 +1,0 @@
-// Package app implements agent application which collects and exports metrics
-// and server which receives and stories the exported metrics.
-package app
